# only consider words that contain less than 3 vowels