Reduces the words by roughly 1k to 7244
A,E,I,O,U,Y are regarded as vowels

if a word has more than 2 vowels, we would need a word with 5 unique characters that doesn't have any vowels in one of the other 5 words

## Benchmarks
`hyperfine --warmup 3 --runs 100 .\wordle.exe`

Before changes:
```
Benchmark 1: .\wordle.exe
  Time (mean ± σ):     152.6 ms ±   7.7 ms    [User: 2.0 ms, System: 42.8 ms]
  Range (min … max):   135.8 ms … 173.1 ms    100 runs
```
After changes:
```
Benchmark 1: .\wordle.exe
  Time (mean ± σ):     118.6 ms ±   6.2 ms    [User: 2.1 ms, System: 36.7 ms]
  Range (min … max):   111.2 ms … 137.2 ms    100 runs
```